requirements.txt
Django==1.11.4
django-mathfilters==0.4.0
djangorestframework==3.6.3
mysqlclient==1.3.10
pkg-resources==0.0.0
pytz==2017.2
requests==2.18.3
djangorestframework-jwt==1.11.0
